Python module isolation: Add runtime Python version
The proper API to get the Python runtime's version was added in
Python 3.11. To get it on earlier versions, parse `sys.version_info`.

Note that this is not added to `modstate` (it is process-wide) as it
should be usable in deinitializers, which can run (just) after the module
is garbage-collected.